chore: add main for omni
