+++ components/src/dynamo/vllm/backend_args.py
@@ -146,153 +146,6 @@ class DynamoVllmArgGroup(ArgGroup):
             choices=[m.value for m in EmbeddingTransferMode],
         )
 
-        # vLLM-Omni
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni",
-            env_var="DYN_VLLM_OMNI",
-            default=False,
-            help="Run as vLLM-Omni worker for multi-stage pipelines (supports text-to-text, text-to-image, etc.).",
-        )
-        add_argument(
-            g,
-            flag_name="--stage-configs-path",
-            env_var="DYN_VLLM_STAGE_CONFIGS_PATH",
-            default=None,
-            help="Path to vLLM-Omni stage configuration YAML file for --omni mode (optional).",
-        )
-
-        # Video encoding
-        add_argument(
-            g,
-            flag_name="--default-video-fps",
-            env_var="DYN_VLLM_DEFAULT_VIDEO_FPS",
-            default=16,
-            arg_type=int,
-            help="Default frames per second for generated videos.",
-        )
-
-        # Diffusion engine-level args (passed to AsyncOmni constructor).
-        # All flags use the --omni- prefix to avoid collisions with vLLM's
-        # native engine flags (e.g. --enforce-eager), which are parsed by a
-        # separate argparse pass and would otherwise be silently consumed here.
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-enable-layerwise-offload",
-            env_var="DYN_VLLM_ENABLE_LAYERWISE_OFFLOAD",
-            default=False,
-            help="Enable layerwise (blockwise) offloading on DiT modules to reduce GPU memory.",
-        )
-        add_argument(
-            g,
-            flag_name="--omni-layerwise-num-gpu-layers",
-            env_var="DYN_VLLM_LAYERWISE_NUM_GPU_LAYERS",
-            default=1,
-            arg_type=int,
-            help="Number of ready layers (blocks) to keep on GPU during generation.",
-        )
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-vae-use-slicing",
-            env_var="DYN_VLLM_VAE_USE_SLICING",
-            default=False,
-            help="Enable VAE slicing for memory optimization in diffusion models.",
-        )
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-vae-use-tiling",
-            env_var="DYN_VLLM_VAE_USE_TILING",
-            default=False,
-            help="Enable VAE tiling for memory optimization in diffusion models.",
-        )
-        add_argument(
-            g,
-            flag_name="--omni-boundary-ratio",
-            env_var="DYN_VLLM_BOUNDARY_RATIO",
-            default=0.875,
-            arg_type=float,
-            help=(
-                "Boundary split ratio for low/high DiT transformers. "
-                "Default 0.875 uses both transformers for best quality. "
-                "Set to 1.0 to load only the low-noise transformer (saves memory). "
-                "Only used with --omni."
-            ),
-        )
-        add_argument(
-            g,
-            flag_name="--omni-flow-shift",
-            env_var="DYN_VLLM_FLOW_SHIFT",
-            default=None,
-            arg_type=float,
-            help="Scheduler flow_shift parameter (5.0 for 720p, 12.0 for 480p). Only used with --omni.",
-        )
-        add_argument(
-            g,
-            flag_name="--omni-diffusion-cache-backend",
-            env_var="DYN_VLLM_DIFFUSION_CACHE_BACKEND",
-            default=None,
-            choices=["cache_dit", "tea_cache"],
-            help=(
-                "Cache backend for diffusion acceleration. "
-                "'cache_dit' enables DBCache + SCM + TaylorSeer. "
-                "'tea_cache' enables TeaCache. Only used with --omni."
-            ),
-        )
-        add_argument(
-            g,
-            flag_name="--omni-diffusion-cache-config",
-            env_var="DYN_VLLM_DIFFUSION_CACHE_CONFIG",
-            default=None,
-            help="Cache configuration as JSON string (overrides defaults). Only used with --omni.",
-        )
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-enable-cache-dit-summary",
-            env_var="DYN_VLLM_ENABLE_CACHE_DIT_SUMMARY",
-            default=False,
-            help="Enable cache-dit summary logging after diffusion forward passes.",
-        )
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-enable-cpu-offload",
-            env_var="DYN_VLLM_ENABLE_CPU_OFFLOAD",
-            default=False,
-            help="Enable CPU offloading for diffusion models to reduce GPU memory usage.",
-        )
-        add_negatable_bool_argument(
-            g,
-            flag_name="--omni-enforce-eager",
-            env_var="DYN_VLLM_ENFORCE_EAGER",
-            default=False,
-            help="Disable torch.compile and force eager execution for diffusion models.",
-        )
-        # Diffusion parallel configuration
-        add_argument(
-            g,
-            flag_name="--omni-ulysses-degree",
-            env_var="DYN_VLLM_ULYSSES_DEGREE",
-            default=1,
-            arg_type=int,
-            help="Number of GPUs used for Ulysses sequence parallelism in diffusion.",
-        )
-        add_argument(
-            g,
-            flag_name="--omni-ring-degree",
-            env_var="DYN_VLLM_RING_DEGREE",
-            default=1,
-            arg_type=int,
-            help="Number of GPUs used for ring sequence parallelism in diffusion.",
-        )
-        add_argument(
-            g,
-            flag_name="--omni-cfg-parallel-size",
-            env_var="DYN_VLLM_CFG_PARALLEL_SIZE",
-            default=1,
-            arg_type=int,
-            choices=[1, 2],
-            help="Number of GPUs used for classifier free guidance parallelism.",
-        )
-
         # Headless mode for multi-node TP/PP
         add_negatable_bool_argument(
             g,
@@ -339,33 +192,6 @@ class DynamoVllmConfig(ConfigBase):
         str, EmbeddingTransferMode
     ]  # resolved to enum in validate()
 
-    # vLLM-Omni
-    omni: bool
-    stage_configs_path: Optional[str] = None
-
-    # Video encoding
-    default_video_fps: int = 16
-
-    # Diffusion engine-level parameters (passed to AsyncOmni constructor).
-    # Field names use omni_ prefix to match the --omni-* CLI flags and avoid
-    # collisions with vLLM's native engine args (e.g. enforce_eager).
-    omni_enable_layerwise_offload: bool = False
-    omni_layerwise_num_gpu_layers: int = 1
-    omni_vae_use_slicing: bool = False
-    omni_vae_use_tiling: bool = False
-    omni_boundary_ratio: float = 0.875
-    omni_flow_shift: Optional[float] = None
-    omni_diffusion_cache_backend: Optional[str] = None
-    omni_diffusion_cache_config: Optional[str] = None
-    omni_enable_cache_dit_summary: bool = False
-    omni_enable_cpu_offload: bool = False
-    omni_enforce_eager: bool = False
-
-    # Diffusion parallel configuration
-    omni_ulysses_degree: int = 1
-    omni_ring_degree: int = 1
-    omni_cfg_parallel_size: int = 1
-
     # Headless mode for multi-node TP/PP
     headless: bool = False
 
@@ -378,7 +204,6 @@ class DynamoVllmConfig(ConfigBase):
         self._resolve_embedding_transfer_mode()
         self._validate_multimodal_role_exclusivity()
         self._validate_multimodal_requires_flag()
-        self._validate_omni_stage_config()
 
     def _resolve_embedding_transfer_mode(self) -> None:
         """Resolve embedding_transfer_mode from string to enum."""
@@ -466,12 +291,4 @@ class DynamoVllmConfig(ConfigBase):
         if self._count_multimodal_roles() == 1 and not self.enable_multimodal:
             raise ValueError(
                 "Use --enable-multimodal when enabling any multimodal component"
-            )
-
-    def _validate_omni_stage_config(self) -> None:
-        """Require stage_configs_path when using --omni."""
-        if self.stage_configs_path and not self.omni:
-            raise ValueError(
-                "--stage-configs-path is only allowed when using --omni. "
-                "Specify a YAML file containing stage configurations for the multi-stage pipeline."
             )
